removed the file I just put in and moved all relevant code to the existing ArchiveCanvas.jsx file
